Remove platform-specific code from core for rendering
